nonzero

+++ fsync.sh
@@ -1,6 +1,9 @@
 fsync /sd/conf/*
 fsync /sd/logs/*
-cp /sd/conf/*.db /homekameleon/conf
-cp /sd/conf/ios.json /homekameleon/conf
-fsync /homekameleon/conf
+if [ -s /sd/conf/ios.json ]
+then
+	cp /sd/conf/*.db /homekameleon/conf
+	cp /sd/conf/ios.json /homekameleon/conf
+	fsync /homekameleon/conf
+fi
 touch /tmp/homekameleon.fsync
